move my test into now existing test file
